refactor: extract scheduler.go into digest, maintenance, and processing files
Split scheduler.go (1575 -> 947 lines) into focused files:
- digest.go: processDigests, flushAndSendDigest, processDailyDigests, sendDailyDigest
- maintenance.go: pruneOldData, processExpiredPremium, retryPending
- processing.go: processSearchListings, scoreAndRecordListings, buildNotifications,
  enrichWithBasePrice, deliverResults, and their helper functions

Also removes dead code: maskPhone (never called outside its own test).

All functions remain methods on *Scheduler in package scheduler.
No signatures or behavior changed.
